+++ dags/web_preaggregated_daily.py
@@ -1,49 +1,29 @@
 from datetime import datetime, UTC, timedelta
 from collections.abc import Callable
-import structlog
+import os
 
 import dagster
-from dagster import DailyPartitionsDefinition, RetryPolicy, Backoff, Jitter, BackfillPolicy
-from clickhouse_driver import Client
-from dags import web_preaggregated_hourly
+from dagster import DailyPartitionsDefinition, BackfillPolicy
 from dags.common import JobOwners
 from dags.web_preaggregated_utils import (
     TEAM_IDS_WITH_WEB_PREAGGREGATED_ENABLED,
     CLICKHOUSE_SETTINGS,
     merge_clickhouse_settings,
     WEB_ANALYTICS_CONFIG_SCHEMA,
+    web_analytics_retry_policy_def,
 )
 from posthog.clickhouse.client import sync_execute
 
 from posthog.models.web_preaggregated.sql import (
-    DISTRIBUTED_WEB_BOUNCES_DAILY_SQL,
-    WEB_BOUNCES_COMBINED_VIEW_SQL,
-    WEB_BOUNCES_DAILY_SQL,
     WEB_BOUNCES_INSERT_SQL,
-    WEB_STATS_COMBINED_VIEW_SQL,
-    WEB_STATS_DAILY_SQL,
-    DISTRIBUTED_WEB_STATS_DAILY_SQL,
     WEB_STATS_INSERT_SQL,
 )
-from posthog.clickhouse.cluster import ClickhouseCluster
-
-logger = structlog.get_logger(__name__)
 
 
 partition_def = DailyPartitionsDefinition(start_date="2020-01-01")
 
-retry_policy_def = RetryPolicy(
-    max_retries=3,
-    delay=60,
-    backoff=Backoff.EXPONENTIAL,
-    jitter=Jitter.PLUS_MINUS,
-)
-
-backfill_policy_def = BackfillPolicy(max_partitions_per_run=14)
-
-
-# Backfill policy to process partitions in groups of 14 days
-web_analytics_backfill_policy = BackfillPolicy.multi_run(max_partitions_per_run=14)
+max_partitions_per_run = int(os.getenv("DAGSTER_WEB_PREAGGREGATED_MAX_PARTITIONS_PER_RUN", 14))
+backfill_policy_def = BackfillPolicy.multi_run(max_partitions_per_run=max_partitions_per_run)
 
 
 def pre_aggregate_web_analytics_data(
@@ -62,17 +42,16 @@ def pre_aggregate_web_analytics_data(
     config = context.op_config
     team_ids = config.get("team_ids", TEAM_IDS_WITH_WEB_PREAGGREGATED_ENABLED)
     extra_settings = config.get("extra_clickhouse_settings", "")
-
     ch_settings = merge_clickhouse_settings(CLICKHOUSE_SETTINGS, extra_settings)
+
+    if not context.partition_time_window:
+        raise dagster.Failure("This asset should only be run with a partition_time_window")
 
     context.log.info(f"Getting ready to pre-aggregate {table_name} for {context.partition_time_window}")
 
-    if context.partition_time_window:
-        start_datetime, end_datetime = context.partition_time_window
-        date_start = start_datetime.strftime("%Y-%m-%d")
-        date_end = end_datetime.strftime("%Y-%m-%d")
-    else:
-        raise dagster.Failure("This asset should only be run with a partition_time_window")
+    start_datetime, end_datetime = context.partition_time_window
+    date_start = start_datetime.strftime("%Y-%m-%d")
+    date_end = end_datetime.strftime("%Y-%m-%d")
 
     try:
         insert_query = sql_generator(
@@ -93,47 +72,6 @@ def pre_aggregate_web_analytics_data(
 
 
 @dagster.asset(
-    name="web_analytics_preaggregated_tables",
-    group_name="web_analytics",
-    retry_policy=retry_policy_def,
-    tags={"owner": JobOwners.TEAM_WEB_ANALYTICS.value},
-)
-def web_analytics_preaggregated_tables(
-    cluster: dagster.ResourceParam[ClickhouseCluster],
-) -> bool:
-    """
-    Create web analytics pre-aggregated tables on all ClickHouse hosts.
-
-    This asset creates both local and distributed tables for web analytics.
-    """
-
-    def drop_tables(client: Client):
-        try:
-            client.execute("DROP TABLE IF EXISTS web_stats_daily SYNC")
-            client.execute("DROP TABLE IF EXISTS web_bounces_daily SYNC")
-            logger.info("dropped_existing_tables", host=client.connection.host)
-        except Exception as e:
-            logger.exception("failed_to_drop_tables", host=client.connection.host, error=str(e))
-            raise
-
-    def create_tables(client: Client):
-        client.execute(WEB_STATS_DAILY_SQL(table_name="web_stats_daily"))
-        client.execute(WEB_BOUNCES_DAILY_SQL(table_name="web_bounces_daily"))
-
-        client.execute(DISTRIBUTED_WEB_STATS_DAILY_SQL())
-        client.execute(DISTRIBUTED_WEB_BOUNCES_DAILY_SQL())
-
-    try:
-        cluster.map_all_hosts(drop_tables).result()
-        cluster.map_all_hosts(create_tables).result()
-        logger.info("web_analytics_tables_setup_completed")
-        return True
-    except Exception as e:
-        logger.exception("web_analytics_tables_setup_failed", error=str(e))
-        raise dagster.Failure(f"Failed to setup web analytics tables: {str(e)}") from e
-
-
-@dagster.asset(
     name="web_analytics_bounces_daily",
     group_name="web_analytics",
     config_schema=WEB_ANALYTICS_CONFIG_SCHEMA,
@@ -142,7 +80,7 @@ def web_analytics_preaggregated_tables(
     backfill_policy=backfill_policy_def,
     metadata={"table": "web_bounces_daily"},
     tags={"owner": JobOwners.TEAM_WEB_ANALYTICS.value},
-    retry_policy=retry_policy_def,
+    retry_policy=web_analytics_retry_policy_def,
 )
 def web_bounces_daily(
     context: dagster.AssetExecutionContext,
@@ -169,7 +107,7 @@ def web_bounces_daily(
     backfill_policy=backfill_policy_def,
     metadata={"table": "web_stats_daily"},
     tags={"owner": JobOwners.TEAM_WEB_ANALYTICS.value},
-    retry_policy=retry_policy_def,
+    retry_policy=web_analytics_retry_policy_def,
 )
 def web_stats_daily(context: dagster.AssetExecutionContext) -> None:
     """
@@ -183,23 +121,6 @@ def web_stats_daily(context: dagster.AssetExecutionContext) -> None:
         table_name="web_stats_daily",
         sql_generator=WEB_STATS_INSERT_SQL,
     )
-
-@dagster.asset(
-    name="web_analytics_combined_views",
-    group_name="web_analytics",
-    description="Creates combined views that automatically merge daily and hourly data using toStartOfDay(now()) boundary.",
-    deps=["web_analytics_preaggregated_tables", web_preaggregated_hourly.web_analytics_preaggregated_hourly_tables],
-    tags={"owner": JobOwners.TEAM_WEB_ANALYTICS.value},
-)
-def web_analytics_combined_views(
-    cluster: dagster.ResourceParam[ClickhouseCluster],
-) -> bool:
-    def create_views(client: Client):
-        client.execute(WEB_STATS_COMBINED_VIEW_SQL())
-        client.execute(WEB_BOUNCES_COMBINED_VIEW_SQL())
-
-    cluster.map_all_hosts(create_views).result()
-    return True
 
 
 # Daily incremental job with asset-level concurrency control
@@ -216,12 +137,6 @@ web_pre_aggregate_daily_job = dagster.define_asset_job(
             }
         }
     },
-)
-
-web_analytics_table_setup_job = dagster.define_asset_job(
-    name="web_analytics_table_setup",
-    selection=["web_analytics_preaggregated_tables"],
-    tags={"owner": JobOwners.TEAM_WEB_ANALYTICS.value},
 )
