+++ dags/web_preaggregated_utils.py
@@ -1,8 +1,15 @@
 from typing import Optional
-from dagster import Field, Array
+from dagster import Backoff, Field, Array, Jitter, RetryPolicy
 
 # TODO: Remove this once we're fully rolled out but this is better than defaulting to all teams
 TEAM_IDS_WITH_WEB_PREAGGREGATED_ENABLED = [1, 2, 55348, 47074]
+
+web_analytics_retry_policy_def = RetryPolicy(
+    max_retries=3,
+    delay=60,
+    backoff=Backoff.EXPONENTIAL,
+    jitter=Jitter.PLUS_MINUS,
+)
 
 # Shared ClickHouse settings for web analytics pre-aggregation
 CLICKHOUSE_SETTINGS = {
